Extract SetupSuccessContent component from SetupSuccess.
